3dPrinterQueue/requirements.txt
flask==3.0.0
bambulabs-api>=2.0.0
python-dotenv==1.0.0
